Update aloe-promptifier.py
##UPDATE

You may now choose to lock if an addition will append to the negative prompt with "neg" or to the positive prompt with "pos". This is useful, if you want a specific trigger keyword in either side of the prompt to automatically append a string.

Example:

If you have you want to negate brightness when you prompt for darkness. You're gonna put brightness type keywords or extra networks to the addition, then in the triggers you can put "dark" or whatever. Then select [NEG] before saving.

Which will make sure that if you type "dark" in the positive prompt, the addition will be added to the negative prompt ! And vice-versa. 

In coming:
- Editing the json right from the webui
- Toggling the extension on and off
- Toggling specific additions on and off
- Extra networks automatic loader/selector to make it easier to add them.