Update aloe-promptifier.py
## Safe Regulator

If you host a Stable Diffusion bot, you can use the regulated.json to type safe words that will in turn make it impossible for people to use a certain amount of words based on your safewords. 

Example:

{
    "chibi|chib|kid|child|young|girl": ["nsfw", "erotica", "nudity"]
}

Add any number of safe-words you want in the first section, separated by vertical bars | and in the second section, you can add the number of banned words you want. If safe word in prompt = banned words don't make it to generation.

## Enable/Disable checkbox

## Make sure to save your .json files when updating. Make copies, just to be sure !